+++ packages/semantic-ui/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -92,9 +92,13 @@ export default function BaseInputTemplate<
       {Array.isArray(schema.examples) && (
         <datalist id={examplesId(id)}>
           {(schema.examples as string[])
-            .concat(schema.default && !schema.examples.includes(schema.default) ? ([schema.default] as string[]) : [])
-            .map((example) => {
-              return <option key={example} value={example} />;
+            .concat(
+              schema.default !== undefined && !schema.examples.map(String).includes(String(schema.default))
+                ? ([schema.default] as string[])
+                : [],
+            )
+            .map((example: any) => {
+              return <option key={String(example)} value={example} />;
             })}
         </datalist>
       )}